feat(auth): implement user login functionality
The Login button on the navbar takes the user to the login page. On form submission the Login form invokes the server action login() which makes a POST request to the API endpoint on the Nest.js server with the credentials. On successful login the server sends a JWT token and the token is saved as a HTTP-only cookie in the client side.